refactor(eval): split the capability probes on what changes together
The 500-line lint on _capability_probes.py was a symptom, and the first
response to it treated it as arithmetic. This addresses what it was pointing
at.

Two defects the line count could not see, both introduced by that first split:

An __all__ block re-exported every moved name back through
_capability_probes. Fifteen names, zero consumers in the repository. Splitting
a module and then re-exporting its moved API keeps the coupling and adds a
name, so no caller ever crosses the new boundary. Deleted.

_subagent_event_kind was written twice, once for the support probe and once
for the concurrency walk. Two definitions of the same domain fact: a runtime
spelling its events sub_agent or child.start would have been taught to one
reader and not the other, and the two probes would have disagreed about the
same stream without either failing. One definition now.

That duplication named the real seam. The support probe and the concurrency
walk share the subagent event vocabulary and change together when a harness
changes how it reports children; override probing does not share it and reads
values through _capability_evidence instead. So the boundary is topology
versus override, not reading versus running, and _capability_topology.py owns
the vocabulary, the lifecycle reader, the request counter, and the peak walk.
Tests reach it directly rather than through _capability_probes, so the
boundary is visible in the tests too.

_capability_probes.py is 396 lines, down from 600 before any of this. That
number is a consequence of deleting dead surface, collapsing a duplicate, and
grouping by reason to change, not the goal it was the first time.

All seven guard mutations still kill a named test and the inverted control
still survives, re-run against the new layout.

Refs #5423
