+++ tests/eval/_harness_capability_test_support.py
@@ -16,6 +16,7 @@ if _path_added:
     sys.path.insert(0, str(EVAL_DIR))
 try:
     import _capability_probes as probes
+    import _capability_topology as topology
     import _harness_capability as capability
 
     _spec = importlib.util.spec_from_file_location("eval_harness_capability", CLI_SCRIPT)
@@ -27,4 +28,4 @@ finally:
     if _path_added:
         sys.path.remove(str(EVAL_DIR))
 
-__all__ = ["MATRIX", "capability", "cli", "probes"]
+__all__ = ["MATRIX", "capability", "cli", "probes", "topology"]
